back-end/requirements.txt
bs4
fastapi
langchain
langchain_community
langchain_core
langchain_groq
langchain_huggingface
langchain_ollama
langchain_openai
langchain_pinecone
langchain_pymupdf4llm
langchain_text_splitters
langdetect
newspaper3k
nltk
pandas
pinecone
pydantic
python-dotenv
requests
streamlit
trafilatura
uvicorn
python-multipart
sentence-transformers
scrapy
decisionrulespy